fix(status): Use in or string serde conversion for status entity
